Fix a crash PaintWorkletImageProvider::GetPaintRecordResult
The root cause of this crash is:
At DiscardableImageMap::GatherDiscardableImages, we got early
exist when the PaintOp::ComputePaintRect. The early out makes
sense because there is nothing to paint if the paint rect is
empty.
Due to the above early out, the |paint_worklet_records_| in the
PictureLayerImpl is empty. But later on at
DrawImageRectOp::RasterWithFlags, we do not have the same
check using PaintOp::ComputePaintRect, so it will try to look
for the RasterContent in the |paint_worklet_records_|.

The solution is that at DrawImageRectOp::RasterWithFlags, if
the PaintOp::ComputePaintRect is empty, we should just early
return because there is nothing to paint, so no paint record
at all.

Bug: 1095163
Change-Id: Ic46636cfe7c256f58261430e07c9ff3a9efa6c80