[Import Maps] Add tests for non-ASCII characters
As the non-bare keys and resolution results of
specifier maps are normalized (i.e. serialized URLs),
raw and percent-encoded non-ASCII strings are
considered equal (as both of them are
percent-encoded after serialization).

This CL tests that raw and percent-encoded versions of
non-ASCII characters are:

- Considered equal in non-bare specifier keys,
- Considered equal in checking backtracking, and
- Not considered equal in bare specifier keys.

For consistency, this CL also adds some backtracking tests
that don't involve non-ASCII characters but use
the same patterns of added non-ASCII tests.

Bug: 848607, https://github.com/WICG/import-maps/issues/207
Change-Id: I6c8ae478da2fcbe0f6df03a608dabdad60644ea1