[Trusted Types] Use sample script value that is actually a script.
Fix-It week: The test passes, but will then report an error in the
test harness, because the string chosen for the example test is not
actually a valid script. If we simply return "0" - which is valid JS -
the bug disappears. (The test doesn't care about the script anyhow, only
about whether it's accepted as input ot not.)

Change-Id: Ie36c27965494f71b796176d3818aefd41b3067ab