Create compositor animation for background-color
This CL is the first step of the background-color animation using
off-thread paint worklet infrastructure. The design doc is here:
https://docs.google.com/document/d/1usCnwWs8HsH5FU_185q6MsrZehFmpl5QgbbB4pvHIjI/edit

In particular, this CL just make sure that we are able to create a
compositor animation for background-color. Then the compositor
thread ticks the animation and dispatches the paint job to
the worker thread to paint.

In the actual paint, we do hard code the result, which means
there is no interpolation implementation at this moment. We
will implement the interpolation logic in a coming CL.

Bug: 1139008, 1148369
Change-Id: I9d2aba2106f38f89bce0e97df29fa0d4f30b2c76