Expose VideoEncoder to workers
This CL mirrors the changes landed in
acbd2eae0763947f78244814c2bdd5628778813d, and exposes VideoEncoder to
workers.

The changes in this CL rely on the conclusions from the previous CL.
Notably, there is a need for a synchronous wait when getting the GPU
factories due to some threading issues, and there is no support for
logging from workers.

Bug: 1094169
Change-Id: I2da6cdb2afd7b80e342268cbbfcb00fea64ad37f